+++ backend/substrapp/serializers/__init__.py
@@ -1,8 +1,4 @@
 from .algo import AlgoSerializer
-from .computetask import OrchestratorAggregateTaskSerializer
-from .computetask import OrchestratorCompositeTrainTaskSerializer
-from .computetask import OrchestratorTestTaskSerializer
-from .computetask import OrchestratorTrainTaskSerializer
 from .datamanager import DataManagerSerializer
 from .datasample import DataSampleSerializer
 from .metric import MetricSerializer
@@ -12,8 +8,4 @@ __all__ = [
     "DataManagerSerializer",
     "MetricSerializer",
     "AlgoSerializer",
-    "OrchestratorTrainTaskSerializer",
-    "OrchestratorTestTaskSerializer",
-    "OrchestratorCompositeTrainTaskSerializer",
-    "OrchestratorAggregateTaskSerializer",
 ]